# Conflicts:
#	logs/edu.log

+++ src/main/java/uz/sfera/edu_platform/payload/CategoryDTO.java
@@ -19,7 +19,7 @@ public class CategoryDTO {
     @NotBlank(message = "Bush bulmasin")
     private String description;
 
-    private byte duration;
+    private int duration;
 
     private double price;
